lib/__init__.py
import torch
from icecream import install

torch.set_num_threads(1)
install()

from .data import *  # noqa
from .deep import *  # noqa
from .env import *  # noqa
from .metrics import *  # noqa
from .neighbors import *  # noqa
from .sam import *  # noqa
from .util import *  # noqa
from .turbo_optimizer import *  # noqa
