requirements.txt
category-encoders==2.3.0
delu==0.0.13
geopy==2.2.0
loguru==0.6.0
numpy==1.21.4
optuna==2.9.1
pandas==1.3.4
pyarrow==6.0.0
scikit-learn==1.0.1
scipy==1.8.0
setuptools==59.5.0  # for torch.utils.tensorboard
tensorboard==2.11.0
tomli==1.2.2
tomli-w==1.0.0
torchinfo==1.5.3
tqdm==4.62.3
xgboost==1.5.0

# Tools
black
flake8
icecream
isort

# Jupyter
ipywidgets
jupyterlab
jupyterlab_code_formatter
voila

# Visualization
bokeh
colorcet
holoviews
matplotlib
seaborn

# new
git+https://github.com/uber-research/TuRBO.git@master
bayesmark
loss-landscapes
plotly